refactor: SvcNo 관련하여 수정
